# o3as/swagger.yml
swagger: "2.0"
info:
  description: This is the swagger file that goes with our server code
  version: "1.0.0"
  title: Swagger REST o3as
consumes:
  - "application/json"
basePath: "/api"
...
paths:
...
  /get_model_info:
    post:
      operationId: "o3as.api.get_model_info"
      tags:
        - "get_model_info"
      summary: "Returning information about a model"
      description: "Return model info"
      produces:
        - "application/json"
      parameters:
        - name: model
          in: query
          type: string
          description: model name
          default: ecmwf-era-interim
          required: true
      responses:
        200:
          description: "Successfully returned model information"
          schema:
            $ref: '#/definitions/ModelInfo'
...
definitions:
...
  ModelInfo:
    type: object
    properties:
      name:
        type: string
      attrs:
        type: string
      dims:
        type: string
      coords:
        type: string